refactor: change code according to the code review

+++ src/controllers/cards.ts
@@ -1,10 +1,11 @@
 import { NextFunction, Request, Response } from 'express';
+import mongoose from 'mongoose';
 
 import { toggleCardLike } from '../utils';
 
 import Card from '../models/card';
 import { Action } from '../types';
-import { BadRequestError, NotFoundError } from '../errors';
+import { BadRequestError, ForbiddenError, NotFoundError } from '../errors';
 
 export const getCards = (req: Request, res: Response, next: NextFunction) => {
   Card.find({})
@@ -21,7 +22,7 @@ export const createCard = (req: Request, res: Response, next: NextFunction) => {
   })
     .then((card) => res.send(card))
     .catch((error) => {
-      if (error instanceof Error && error.name === 'ValidationError') {
+      if (error instanceof mongoose.Error.ValidationError) {
         return next(new BadRequestError('Переданы некорректные данные'));
       }
       return next(error);
@@ -30,15 +31,19 @@ export const createCard = (req: Request, res: Response, next: NextFunction) => {
 
 export const deleteCard = (req: Request, res: Response, next: NextFunction) => {
   const { cardId } = req.params;
+  const { _id } = req.user;
 
-  Card.findByIdAndDelete(cardId)
-    .orFail(new Error('DocumentNotFoundError'))
-    .then((card) => res.send(card))
-    .catch((error) => {
-      if (error instanceof Error && error.message === 'DocumentNotFoundError') {
-        return next(new NotFoundError('Запрашиваемая информация не найдена'));
+  Card.findById(cardId)
+    .orFail(new NotFoundError('Пост не найден'))
+    .then((card) => {
+      if (card.owner.toString() !== _id) {
+        throw new ForbiddenError('Вы не можете удалить чужой пост');
       }
-      if (error instanceof Error && error.name === 'CastError') {
+      return Card.findByIdAndDelete(cardId);
+    })
+    .then(() => res.send({ message: 'Пост удален' }))
+    .catch((error) => {
+      if (error instanceof mongoose.Error.CastError) {
         return next(new BadRequestError('Переданы некорректные данные'));
       }
       return next(error);
@@ -49,32 +54,12 @@ export const likeCard = (req: Request, res: Response, next: NextFunction) => {
   const { _id } = req.user;
   const { cardId } = req.params;
 
-  toggleCardLike(cardId, _id, Action.LIKE)
-    .then((card) => res.send(card))
-    .catch((error) => {
-      if (error instanceof Error && error.message === 'DocumentNotFoundError') {
-        return next(new NotFoundError('Запрашиваемая информация не найдена'));
-      }
-      if (error instanceof Error && error.name === 'CastError') {
-        return next(new BadRequestError('Переданы некорректные данные'));
-      }
-      return next(error);
-    });
+  toggleCardLike(cardId, _id, Action.LIKE, req, res, next);
 };
 
 export const dislikeCard = (req: Request, res: Response, next: NextFunction) => {
   const { _id } = req.user;
   const { cardId } = req.params;
 
-  toggleCardLike(cardId, _id, Action.DISLIKE)
-    .then((card) => res.send(card))
-    .catch((error) => {
-      if (error instanceof Error && error.message === 'DocumentNotFoundError') {
-        return next(new NotFoundError('Запрашиваемая информация не найдена'));
-      }
-      if (error instanceof Error && error.name === 'CastError') {
-        return next(new BadRequestError('Переданы некорректные данные'));
-      }
-      return next(error);
-    });
+  toggleCardLike(cardId, _id, Action.DISLIKE, req, res, next);
 };
